requirements.txt
mss==9.0.1
pillow==10.0.0
pyautogui==0.9.54
keyboard==0.13.5
pynput==1.7.6